Support Juju 1.25 in store-only mode.
Charms that need to support both Juju 1.25 and Juju 2.0+
cannot declare the snap resource in metadata.yaml. They will
only be able to install and update from the snap store.

hookenv.resource_get() fails hard if called from a Juju 1.25
deploy. We can fix this problem there, but need to work around
it in this layer until that fix is released.