Don't change the snapd refresh timer if not set
If the refresh timer is not explicitly set via config, it should not be
written to the system.  This mainly affects subordinates, where the
subordinate using layer:snap and no explicit refresh value could wipe
out the prinicpal's explicit refresh value due to the refresh timer
being a global setting.